backend/requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
python-dotenv
langchain-google-genai
